Added a big list of stats
This is super-long already, and it only represents the stats on the first page of the character sheet. Waaay more work than I thought.  I'd really like to put more effort to some sort of dynamic system for defining stats.  Some attempts are included in the experiments folder.

This list and its model should be considered highly work-in-progress until the related objects and classes have been well defined and fleshed out.  As it is, a few things already should change, especially attack damage, but I'm out of energy right now.

We should also start to refactor the stats into individual components or directives for display and editing, rather than duplicating hunks of code.  The dynamic lists (eg. personality, equipment, etc.) are prime candidates.